Removed DB Upload

+++ EM6400/dropbox_upload.py
@@ -1,40 +1,34 @@
-from configuration import THRESHOLD_TIME, DATA_BASE_PATH, BASE_UPLOAD_PATH
-from dbupload import DropboxConnection
-import db_password
+from configuration import THRESHOLD_TIME, DATA_BASE_PATH, BASE_UPLOAD_PATH,URL_PATH
 import glob
 import os
 import time
+import requests
 
-email=db_password.email
-password=db_password.password
-dropbox_upload_log=open(DATA_BASE_PATH+"db_log.txt","w")
+file_upload_log=open(DATA_BASE_PATH+"file_upload_log.txt","w")
 
+while True:
 
-folders=os.listdir(DATA_BASE_PATH)
-try:
-    # Create the connection
-    conn = DropboxConnection(email, password)
-    #List of files
-    for folder in folders:
-        list_of_files=glob.glob(str(DATA_BASE_PATH)+str(folder)+str("/*.csv"))
-        for f in list_of_files:    
-            
-           
-            if int(time.time())-int(os.stat(f).st_mtime)>THRESHOLD_TIME:
+    folders=os.listdir(DATA_BASE_PATH)
+    try:
+        
+        for folder in folders:
+            list_of_files=glob.glob(str(DATA_BASE_PATH)+str(folder)+str("/*.csv"))
+            for f in list_of_files:    
+                if int(time.time())-int(os.stat(f).st_mtime)>THRESHOLD_TIME:
                 
-                dropbox_upload_log=open(DATA_BASE_PATH+"db_log.txt","w")
-                dropbox_upload_log.write(str(f) +" will be uploaded")
-                dropbox_upload_log.close()
-                r=open(DATA_BASE_PATH+"db_log.txt","r")
-                a=r.read()
-                r.close()
-                
-                conn.upload_file(f,BASE_UPLOAD_PATH+str(folder),f)
-                os.remove(f)
-except:
-    dropbox_upload_log=open(DATA_BASE_PATH+"db_log.txt","w")
-    dropbox_upload_log.write("Upload failed"+str(int(time.time())))
-    dropbox_upload_log.close()
-else:
-    pass
+                    file_upload_log=open(DATA_BASE_PATH+"file_upload_log.txt","w")
+                    file_upload_log.write(str(f) +" will be uploaded")
+                    file_upload_log.close()
+                    r=open(f,"r")
+                    files = {'myfile': r} 
+                    g = requests.post(url=URL_PATH,files=files)
+                    if g.status_code==200:
+                        os.remove(f)
+    except:
+        file_upload_log=open(DATA_BASE_PATH+"file_upload_log.txt","w")
+        file_upload_log.write("Upload failed"+str(int(time.time())))
+        file_upload_log.close()
+    else:
+        pass
+    time.sleep(THRESHOLD_TIME)
